+++ NEWS.md
@@ -4,6 +4,7 @@ Breaking changes
 * `names_section` parameter is replaced by a named vector in `pattern`
 
 Other changes
+* Add parameter `pattern.table` to allow for user specific pattern titles
 * `create_vignette_last_modif()` creates a vignette with dates of files modification in git
 * Added summary tables for each section
 * Added a `NEWS.md` file to track changes to the package.